Renamed translations folder to list folder
