+++ src/containers/List/ListWrapper.tsx
@@ -4,9 +4,9 @@ import { Loading, FilterList } from '../../components'
 import { useRouter } from '../../utils/hooks/useRouter'
 import useListApplications from '../../utils/hooks/useListApplications'
 import strings from '../../utils/constants'
-import getDefaultUserRole from '../../utils/helpers/translations/findUserRole'
+import getDefaultUserRole from '../../utils/helpers/list/findUserRole'
 import { useUserState } from '../../contexts/UserState'
-import mapColumnsByRole from '../../utils/helpers/translations/mapColumnsByRole'
+import mapColumnsByRole from '../../utils/helpers/list/mapColumnsByRole'
 import { ColumnDetails } from '../../utils/types'
 import { USER_ROLES } from '../../utils/data'
 import { Link } from 'react-router-dom'
