--- src/utils/graphql/queries/getApplicationsStages.query.ts
@@ -1,14 +1,0 @@
-import { gql } from '@apollo/client'
-
-export default gql`
-  query getApplicationsStages($serials: [String!]) {
-    applicationStageStatusAlls(
-      filter: { serial: { in: $serials } }
-      condition: { stageIsCurrent: true }
-    ) {
-      nodes {
-        ...Stage
-      }
-    }
-  }
-`
