+++ src/utils/helpers/list/findUserRole.ts
